pyproject.toml
[build-system]
requires = ["scikit-build-core[pyproject]>=0.9.2"]
build-backend = "scikit_build_core.build"

[project]
name = "nekko_api"
dynamic = ["version"]
description = "OpenAI API compatible llama.cpp server"
readme = "README.md"
license = { text = "Apache License 2.0" }
authors = [
]
dependencies = [
    "typing-extensions>=4.5.0",
    "numpy>=1.20.0",
    "diskcache>=5.6.1",
    "jinja2>=2.11.3",
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]


[project.optional-dependencies]
server = [
    "uvicorn>=0.22.0",
    "fastapi>=0.100.0",
    "pydantic-settings>=2.0.1",
    "sse-starlette>=1.6.1",
    "starlette-context>=0.3.6,<0.4",
    "PyYAML>=5.1",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp>=1.29.0",
]
test = [
    "openai>=1.57.2",
    "pytest>=7.4.0",
    "httpx>=0.24.1",
    "scipy>=1.10",
    "fastapi>=0.100.0",
    "sse-starlette>=1.6.1",
    "starlette-context>=0.3.6,<0.4",
    "pydantic-settings>=2.0.1",
    "huggingface-hub>=0.23.0"
]
dev = [
    "mypy>=1.13.0",
    "black>=23.3.0",
    "twine>=4.0.2",
    "mkdocs>=1.4.3",
    "mkdocstrings[python]>=0.22.0",
    "mkdocs-material>=9.1.18",
    "pytest>=7.4.0",
    "httpx>=0.24.1",
]
all = [
    "nekko_api[server,test,dev]",
]

[tool.scikit-build]
wheel.packages = ["llama_cpp"]
cmake.verbose = true
cmake.minimum-version = "3.21"
minimum-version = "0.5.1"
sdist.include = [".git", "vendor/llama.cpp/*"]

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
input = "llama_cpp/__init__.py"

[project.urls]
Homepage = "https://github.com/aifoundry-org/NekkoAPI"
Issues = "https://github.com/aifoundry-org/NekkoAPI/issues"
Documentation = "https://github.com/aifoundry-org/NekkoAPI/wiki"

[tool.pytest.ini_options]
testpaths = "tests"
